+++ layers/Engine/packages/engine/src/Schema/SchemaDefinitionService.php
@@ -5,17 +5,25 @@ declare(strict_types=1);
 namespace PoP\Engine\Schema;
 
 use PoP\Engine\TypeResolvers\RootTypeResolver;
-use PoP\ComponentModel\Facades\Instances\InstanceManagerFacade;
 use PoP\ComponentModel\Schema\SchemaDefinitionService as ComponentModelSchemaDefinitionService;
+use PoP\ComponentModel\Instances\InstanceManagerInterface;
 
 class SchemaDefinitionService extends ComponentModelSchemaDefinitionService implements SchemaDefinitionServiceInterface
 {
+    public function __construct(protected InstanceManagerInterface $instanceManager)
+    {
+    }
+
+    public function getTypeResolverTypeSchemaKey(string $typeResolverClass): string
+    {
+        $typeResolver = $this->instanceManager->getInstance($typeResolverClass);
+        return $this->getTypeSchemaKey($typeResolver);
+    }
+
     public function getRootTypeSchemaKey(): string
     {
-        $instanceManager = InstanceManagerFacade::getInstance();
         $rootTypeResolverClass = $this->getRootTypeResolverClass();
-        $rootTypeResolver = $instanceManager->getInstance($rootTypeResolverClass);
-        return $this->getTypeSchemaKey($rootTypeResolver);
+        return $this->getTypeResolverTypeSchemaKey($rootTypeResolverClass);
     }
 
     public function getRootTypeResolverClass(): string