+++ layers/GraphQLByPoP/packages/graphql-server/src/ComponentConfiguration.php
@@ -27,6 +27,7 @@ class ComponentConfiguration
     private static bool $enableSettingMutationSchemeByURLParam = false;
     private static bool $enableEnablingGraphQLIntrospectionByURLParam = false;
     private static bool $addGraphQLIntrospectionPersistedQuery = false;
+    private static bool $addConnectionFromRootToQueryRootAndMutationRoot = false;
 
     public static function addSelfFieldForRootTypeToSchema(): bool
     {
@@ -303,6 +304,24 @@ class ComponentConfiguration
         // Define properties
         $envVariable = Environment::ADD_GRAPHQL_INTROSPECTION_PERSISTED_QUERY;
         $selfProperty = &self::$addGraphQLIntrospectionPersistedQuery;
+        $defaultValue = false;
+        $callback = [EnvironmentValueHelpers::class, 'toBool'];
+
+        // Initialize property from the environment/hook
+        self::maybeInitializeConfigurationValue(
+            $envVariable,
+            $selfProperty,
+            $defaultValue,
+            $callback
+        );
+        return $selfProperty;
+    }
+
+    public static function addConnectionFromRootToQueryRootAndMutationRoot(): bool
+    {
+        // Define properties
+        $envVariable = Environment::ADD_CONNECTION_FROM_ROOT_TO_QUERYROOT_AND_MUTATIONROOT;
+        $selfProperty = &self::$addConnectionFromRootToQueryRootAndMutationRoot;
         $defaultValue = false;
         $callback = [EnvironmentValueHelpers::class, 'toBool'];