+++ layers/Schema/packages/comments/src/TypeResolvers/CommentTypeResolver.php
@@ -45,7 +45,7 @@ class CommentTypeResolver extends AbstractTypeResolver
 
     public function getSchemaTypeDescription(): ?string
     {
-        return $this->translationAPI->__('Comments added to posts', 'comments');
+        return $this->translationAPI->__('Comments added to custom posts', 'comments');
     }
 
     public function getID(object $resultItem): string | int | null